restructure some files and folders
