+++ src/components/cart/cart-dropdown/CartDropdown.jsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 
-import CustomButton from '../custom-button/CustomButton';
+import CustomButton from '../../custom-button/CustomButton';
 
-import { selectCartItems } from '../../redux/cart/cart.selectors';
-import { toggleCartHidden } from '../../redux/cart/cart.actions';
+import { selectCartItems } from '../../../redux/cart/cart.selectors';
+import { toggleCartHidden } from '../../../redux/cart/cart.actions';
 
 import './cart-dropdown.scss';
 import CartItem from '../cart-item/CartItem';
