+++ src/components/cart/cart-icon/CartIcon.jsx
