+++ src/components/cart/cart-icon/cart-icon.scss
