--- src/components/collection-overview/collection-overview.scss
@@ -1,4 +1,0 @@
-.collections-overview {
-    display: flex;
-    flex-direction: column;
-}
