--- src/components/collection-preview/collection-preview.scss
@@ -1,15 +1,0 @@
-.collection-preview {
-    display: flex;
-    flex-direction: column;
-    margin-bottom: 30px;
-  
-    .title {
-      font-size: 28px;
-      margin-bottom: 25px;
-    }
-  
-    .preview {
-      display: flex;
-      justify-content: space-between;
-    }
-  }
