+++ src/components/collection/CollectionPreview.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+import './index.scss';
+
+import CollectionItem from './CollectionItem';
+
+const CollectionPreview = ({ title, items }) => (
+  <div className='collection-preview'>
+    <h1 className='title'>{title.toUpperCase()}</h1>
+    <div className='preview'>
+      {items
+        .filter((item, idx) => idx < 4)
+        .map((item) => (
+          <CollectionItem key={item.id} item={item} />
+        ))}
+    </div>
+  </div>
+);
+
+export default CollectionPreview;
