+++ src/components/collection/index.scss
@@ -50,3 +50,8 @@
     }
   }
 }
+
+.collections-overview {
+  display: flex;
+  flex-direction: column;
+}
