+++ src/components/homepage/HomePage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import Directory from '../directory/Directory';
+import Directory from '../../features/directory/Directory';
 import './homepage.scss';
 
 const HomePage = () => {
