--- src/components/sign-up/sign-up.scss
@@ -1,9 +1,0 @@
-.sign-up {
-    display: flex;
-    flex-direction: column;
-    width: 380px;
-
-    .title {
-        margin: 10px 0px;
-    }
-}
