+++ src/features/authentication/Authentication.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
-
-import SignIn from '../../components/sign-in/SignIn';
-import SignUp from '../../components/sign-up/SignUp';
-
+import SignIn from '../../components/authentication/SignIn';
+import SignUp from '../../components/authentication/SignUp';
 import './sign-in-sign-up.scss';
 
 const Authentication = () => (
