+++ src/features/collection/CollectionPage.jsx
@@ -1,9 +1,6 @@
 import React from 'react';
-import { connect } from 'react-redux';
 
-import CollectionItem from '../../components/collection-item/CollectionItem';
-
-import { selectCollection } from '../../redux/shop/shop.selectors';
+import CollectionItem from '../../components/collection/CollectionItem';
 
 import './collection.scss';
 
@@ -21,9 +18,4 @@ const CollectionPage = ({ collection }) => {
   );
 };
 
-const mapStateToProps = (state, ownProps) => ({
-  collection: selectCollection(ownProps.match.params.collectionId)(state),
-});
-
-//export default connect(mapStateToProps)(CollectionPage);
 export default CollectionPage;
